fixed a bug where number of seconds was truncated from the api call duration

+++ shared/selene/data/metrics/entity/api.py
@@ -1,5 +1,6 @@
 from dataclasses import dataclass
 from datetime import datetime
+from decimal import Decimal
 
 
 @dataclass
@@ -7,7 +8,7 @@ class ApiMetric(object):
     url: str
     access_ts: datetime
     api: str
-    duration: int
+    duration: Decimal
     http_status: int
     id: str = None
     account_id: str = None
